Self check protection and enforcing movement rules
I added functionality that goes through the list of all moves a piece can make, and check if making said move would put their own king in check. These moves are then removed from the possible move array and are not displayed on the board. Also I added logic that makes sure pieces can only move to squares that show that moving there is a potential move. Essentially enforcing the rules of chess.